Moved helper functions to a separate file.
